整体优化

+++ treasure-api/src/main/java/io/treasure/entity/MerchantSalesRewardRecordEntity.java
@@ -22,7 +22,7 @@ public class MerchantSalesRewardRecordEntity {
     private String auditStatus;
     private Integer auditComment;
     private Integer withDrawTime;
-    private Integer commissionVolume;
+    private BigDecimal commissionVolume;
 
     private BigDecimal salesVolume;
